requirements.txt
Babel==2.7.0
Django==2.2.3
django-phonenumber-field==3.0.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
phonenumbers==8.10.14
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
